requirements.txt
mkdocs-material
mkdocs-gen-nav-plugin
flask 
authlib
werkzeug
python-dotenv